refactor: adopt new change from step1
