+++ src/main/kotlin/lotto/dto/IssuedTickets.kt
@@ -11,11 +11,11 @@ data class IssuedTickets(private val tickets: List<LottoTicket>) {
     operator fun plus(ticket: LottoTicket): IssuedTickets {
         return IssuedTickets(
             tickets.toMutableList()
-            .also { it.add(ticket) }.toList()
+                .also { it.add(ticket) }.toList()
         )
     }
 
-   operator fun plus(other: IssuedTickets): IssuedTickets {
+    operator fun plus(other: IssuedTickets): IssuedTickets {
         return IssuedTickets(
             this.tickets + other.tickets
         )
